Computer/computer_requirements.txt
tensorflow == 2.9.1
Pillow == 9.2.0
numpy == 1.23.2
opencv-python == 4.6.0.66
matplotlib == 3.5.3
scikit-learn == 1.1.2